Refactor and rename useModal hook
